// src/types/gastoContable.ts
// Generated by https://quicktype.io

export interface GastoContableRegistrarProps {
    fecha_doc:      Date;
    nroComprobante: string;
    total:          number;
    usuario:        string;
    proveedorId:    string;
    gastoId:        string;
}

// Generated by https://quicktype.io

export interface ListarGastoContableProps {
    idGastoContable: string;
    fecha_doc:      Date;
    nroComprobante: string;
    razonSocial:    string;
    ruc:            string;
    total:          number;
    usuario:        string;
    isActive:       boolean;
    createdAt:      Date;
}